Move redux stuff around
